+++ docs/listings.json
@@ -2,13 +2,14 @@
   {
     "listing": "/index.html",
     "items": [
+      "/posts/12_fcc-data/index.html",
       "/posts/10_cori_data_fcc_overview/index.html",
-      "/posts/pg_duckdb/index.html",
-      "/posts/micropolitan_formd/index.html",
-      "/posts/formd_intro/index.html",
-      "/posts/urisa-2023/index.html",
-      "/posts/mapping_rural_tips/index.html",
-      "/posts/awesomejq/index.html"
+      "/posts/06_pg_duckdb/index.html",
+      "/posts/05_micropolitan_formd/index.html",
+      "/posts/04_formd_intro/index.html",
+      "/posts/03_urisa-2023/index.html",
+      "/posts/02_mapping_rural_tips/index.html",
+      "/posts/01_awesomejq/index.html"
     ]
   }
 ]
