+++ docs/posts/01_awesomejq/thumbnail.png
