+++ docs/posts/02_mapping_rural_tips/images/cartogram_dorling_rural_pop.png
