+++ docs/posts/02_mapping_rural_tips/images/cartogram_ncont_pct_rural_pop.png
