+++ docs/posts/02_mapping_rural_tips/images/choropleth_normalized.png
