+++ docs/posts/02_mapping_rural_tips/thumbnail.png
