+++ docs/posts/05_micropolitan_formd/images/top10_vc_micropolitans.png
