+++ docs/posts/06_pg_duckdb/DuckDB_Postgre_SQL_Logo.png
