+++ docs/site_libs/quarto-html/quarto-syntax-highlighting-66ab7fd5e73b7f0a764e0d49b3e29ab1.css
@@ -85,6 +85,7 @@ code span.st {
 
 code span.cf {
   color: #003B4F;
+  font-weight: bold;
   font-style: inherit;
 }
 
@@ -193,6 +194,7 @@ code span.dv {
 
 code span.kw {
   color: #003B4F;
+  font-weight: bold;
   font-style: inherit;
 }
 
@@ -200,4 +202,4 @@ code span.kw {
   content: "</";
 }
 
-/*# sourceMappingURL=debc5d5d77c3f9108843748ff7464032.css.map */
+/*# sourceMappingURL=959b87231af1ebfcdaf83f2a1def939a.css.map */
