+++ index.qmd
@@ -5,8 +5,6 @@ listing:
   sort: "date desc"
   type: default
   categories: false
-  exclude:
-    path: posts/07_bead/proj-fcc-report/index.qmd
   sort-ui: false
   filter-ui: false
 page-layout: full
