+++ package.json
@@ -26,9 +26,10 @@
     "build": "npm run render",
     "clean": "rm -rf _freeze/* _site/*",
     "dev": "quarto render --output-dir _site",
-    "preview": "npm run clean && quarto render --output-dir _site && python -m http.server -d _site 3000",
+    "preview": "npm run clean && quarto render --output-dir _site && npm run start",
     "render": "quarto render --output-dir docs",
     "render:mermaid": "mmdc -i assets/input.mmd -o assets/output.svg",
+    "start": "python -m http.server -d _site 3000",
     "test:mermaid": "mmdc -i assets/input.mmd -o assets/output.svg",
     "mermaid": "mmdc"
   }
